Support Wasm32 build for RPC-client
