+++ account-decoder/Cargo.toml
@@ -25,7 +25,12 @@ solana-sdk = { path = "../sdk", version = "=1.15.0" }
 spl-token = { version = "=3.5.0", features = ["no-entrypoint"] }
 spl-token-2022 = { version = "=0.4.3", features = ["no-entrypoint"] }
 thiserror = "1.0"
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 zstd = "0.11.2"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+zstd = { version = "0.11.2", default-features = false }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
