+++ programs/address-lookup-table/Cargo.toml
@@ -21,7 +21,7 @@ solana-frozen-abi-macro = { path = "../../frozen-abi/macro", version = "=1.15.0"
 solana-program = { path = "../../sdk/program", version = "=1.15.0" }
 thiserror = "1.0"
 
-[target.'cfg(not(target_os = "solana"))'.dependencies]
+[target.'cfg(not(any(target_os = "solana", target_arch = "wasm32")))'.dependencies]
 solana-program-runtime = { path = "../../program-runtime", version = "=1.15.0" }
 solana-sdk = { path = "../../sdk", version = "=1.15.0" }
