+++ programs/config/Cargo.toml
@@ -14,8 +14,10 @@ bincode = "1.3.3"
 chrono = { version = "0.4.22", features = ["serde"] }
 serde = "1.0.144"
 serde_derive = "1.0.103"
-solana-program-runtime = { path = "../../program-runtime", version = "=1.15.0" }
 solana-sdk = { path = "../../sdk", version = "=1.15.0" }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+solana-program-runtime = { path = "../../program-runtime", version = "=1.15.0" }
 
 [dev-dependencies]
 solana-logger = { path = "../../logger", version = "=1.15.0" }
