+++ rpc-client/src/rpc_client.rs
@@ -21,7 +21,6 @@ use {
         nonblocking::{self, rpc_client::get_rpc_request_str},
         rpc_sender::*,
     },
-    serde::Serialize,
     serde_json::Value,
     solana_account_decoder::{
         parse_token::{UiTokenAccount, UiTokenAmount},
@@ -41,10 +40,9 @@ use {
         epoch_schedule::EpochSchedule,
         fee_calculator::{FeeCalculator, FeeRateGovernor},
         hash::Hash,
-        message::{v0, Message as LegacyMessage},
         pubkey::Pubkey,
         signature::Signature,
-        transaction::{self, uses_durable_nonce, Transaction, VersionedTransaction},
+        transaction,
     },
     solana_transaction_status::{
         EncodedConfirmedBlock, EncodedConfirmedTransactionWithStatusMeta, TransactionStatus,
@@ -53,64 +51,10 @@ use {
     std::{net::SocketAddr, str::FromStr, sync::Arc, time::Duration},
 };
 
-#[derive(Default)]
-pub struct RpcClientConfig {
-    pub commitment_config: CommitmentConfig,
-    pub confirm_transaction_initial_timeout: Option<Duration>,
-}
-
-impl RpcClientConfig {
-    pub fn with_commitment(commitment_config: CommitmentConfig) -> Self {
-        RpcClientConfig {
-            commitment_config,
-            ..Self::default()
-        }
-    }
-}
-
-/// Trait used to add support for versioned messages to RPC APIs while
-/// retaining backwards compatibility
-pub trait SerializableMessage: Serialize {}
-impl SerializableMessage for LegacyMessage {}
-impl SerializableMessage for v0::Message {}
-
-/// Trait used to add support for versioned transactions to RPC APIs while
-/// retaining backwards compatibility
-pub trait SerializableTransaction: Serialize {
-    fn get_signature(&self) -> &Signature;
-    fn get_recent_blockhash(&self) -> &Hash;
-    fn uses_durable_nonce(&self) -> bool;
-}
-impl SerializableTransaction for Transaction {
-    fn get_signature(&self) -> &Signature {
-        &self.signatures[0]
-    }
-    fn get_recent_blockhash(&self) -> &Hash {
-        &self.message.recent_blockhash
-    }
-    fn uses_durable_nonce(&self) -> bool {
-        uses_durable_nonce(self).is_some()
-    }
-}
-impl SerializableTransaction for VersionedTransaction {
-    fn get_signature(&self) -> &Signature {
-        &self.signatures[0]
-    }
-    fn get_recent_blockhash(&self) -> &Hash {
-        self.message.recent_blockhash()
-    }
-    fn uses_durable_nonce(&self) -> bool {
-        self.uses_durable_nonce()
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct GetConfirmedSignaturesForAddress2Config {
-    pub before: Option<Signature>,
-    pub until: Option<Signature>,
-    pub limit: Option<usize>,
-    pub commitment: Option<CommitmentConfig>,
-}
+pub use crate::nonblocking::rpc_client::{
+    GetConfirmedSignaturesForAddress2Config, RpcClientConfig, SerializableMessage,
+    SerializableTransaction,
+};
 
 /// A client of a remote Solana node.
 ///
